backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=2.8
drf-nested-routers>=0.93
django-cors-headers>=4.0
Pillow>=9.0
psycopg2-binary>=2.9.9
